// src/pages/Dashboard/MainViewContent/pages/Customer/Interfaces/IDocument.ts
export interface DocumentCustomerData{
  id: number;
  customerId: number;
  fileName: string;
  fileSize?: number;
  dateUpload: string;
  tokenPagination: string;
  Count:number;
}

export interface DocumentBusinessData{
  id: number;
  businessId: number;
  fileName: string;
  fileSize?: number;
  dateUpload: string;
  tokenPagination: string;
  Count:number;
}